feat(app): email base templates
