+++ src/lib/shared/services/transactional/api/welcome-user.tsx
@@ -1,6 +1,6 @@
-import { WelcomeUser } from '../lib/emails/welcome-user';
 import { SendEmailProps } from '../lib/transactional-types';
 import { transactionalEmailService } from '../model';
+import { WelcomeUser } from '../ui/emails/welcome-user';
 
 export const send = (to: SendEmailProps['to']) =>
   transactionalEmailService.send({
